+++ Taller/src/GHAMatrix.c
@@ -8,12 +8,12 @@
 
 
     Entrada:
-        double Mjd_UT1:                   Fecha Juliana modificada
-        double matrizTransormada[3][3]:   Guarda el resultado de la transformacion
-        double (*eop)[13]:                Variable con datos almacenados
+        long double Mjd_UT1:                   Fecha Juliana modificada
+        long double matrizTransormada[3][3]:   Guarda el resultado de la transformacion
+        long double (*eop)[13]:                Variable con datos almacenados
 */
 
-void ghaMatrix ( double Mjd_UT1,  double matrizTransformada[3][3],  double (*eop)[13])
+void ghaMatrix ( long double Mjd_UT1,  long double matrizTransformada[3][3],  long double (*eop)[13])
 {
     for (int i=0; i<3; i++)
     {
@@ -23,6 +23,6 @@ void ghaMatrix ( double Mjd_UT1,  double matrizTransformada[3][3],  double (*eop
         }
     }
 
-     double g = gast(Mjd_UT1, eop);
+     long double g = gast(Mjd_UT1, eop);
     R_z(g, matrizTransformada);
 }
